ingestion/requirements.txt
# Core dependencies
azure-functions>=1.18.0
azure-cosmos>=4.5.1
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
azure-monitor-opentelemetry>=1.2.0

# HTTP and retry logic
requests>=2.31.0
tenacity>=8.2.3

# Data validation
pydantic>=2.5.0

# AI/ML
openai>=1.12.0

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
slowapi>=0.1.9
sse-starlette>=1.8.2

# Caching
redis>=5.0.1

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0
requests-mock>=1.11.0

# Utilities
python-dotenv>=1.0.0
